Allow operation on SH4 CPUs and remove compiler warnings.
Simplify the GET_UNALIGNED_W macro and use it in all cases instead of making it
conditional on CPU types. This missed some CPUs that needed it (e.g. SH4), and
in any case the implementation caused "dereferencing type-punned pointer will
break strict-aliasing rules" warnings.

Enable extra warnings, but disable signed comparison and missing field
initializer warnings as these are not helpful.

Update write_boot_label() so that the boot_sector_16 and boot_sector cases are
handled separately instead of using an aliased pointer, as that causes
"dereferencing type-punned pointer will break strict-aliasing rules" warnings.

Make date_dos2unix(), usage() and cdiv() static functions as they are only used
in the files in which they are declared.

Update bad_name() and lfn_get() so that the extension is processed separately
instead of by indexing past the end of the name field as that causes "array
subscript is above array bounds" warnings.

Update the dosfsck() main function to avoid a warning that free_clusters may
be used uninitialized. Do not print the final count of files and clusters when
dosfsck is run with the "-b" option because the used files and clusters have
not been counted in this case.

Alter the setup_tables() function so that it does not cause an "array subscript
is below array bounds" warning.

Signed-off-by: Daniel Baumann <mail@daniel-baumann.ch>