Support long file names in volume labeling code
The code to find the volume label directory entry in find_volume_de()
did not consider long file names so far. Directory entries that make up
long file names have four attribute bits set, including the "volume"
bit.

This caused the code to mistake a directory entry that is part of a
long file name as the volume name entry. If such an entry is found
first, fatlabel would print garbage when asked to display the label and
mangle the long file name when asked to set it. The latter would lead
to the loss of the long file name and require a fsck to clean up.

Change so that the set of attributes equal that of LFN entries will no
langer match as a volume label.

Signed-off-by: Andreas Bombe <aeb@debian.org>