fsck.fat: Fix read beyond end of array on FAT12
When a FAT12 filesystem contains an odd number of clusters, setting the
last cluster with set_fat() will trigger a read of the next entry,
which does not exist in the fat array allocated for this.

Round up the allocation to an even number of FAT entries for FAT12 so
that this is fixed without introducing special casing in get_fat().

Signed-off-by: Andreas Bombe <aeb@debian.org>