fsck: Verify first cluster of a file is not 1
Previously the FAT was checked for any out of bounds entries which were
cleared. However the first cluster of a file, as specified in its
directory entry, was not verified to not be 1.

In addition to missing this filesystem error, code that correctly
assumes the FAT contained no bad entries anymore could still look up
invalid table indices depending on the value stored in entry 1. With
the right values and FAT size this can lead to a segfault by accessing
unallocated memory.

Now test_file() will ignore files where the first cluster equals 1 and
an additional check in check_file() will truncate them.

This bug was reported in http://bugs.debian.org/773885 by Jakub Wilk.

Signed-off-by: Andreas Bombe <aeb@debian.org>