Convert build system to autoconf/automake
In preparation for fixing the horribly outdated and broken device
checking - which will likely involve using additional libraries like
libblkid - as well as making this package portable to other operating
systems, the build system is now the tried and true autoconf/automake
combination which should make both goals a little more straightforward.

The release version number and date are now in configure.ac and
substituted by configure where they are needed. Now it is no longer
necessary to change the number in multiple places for a release and the
man pages get the number substituted directly into them, making the
update-version.sh script and the VERSION file obsolete.

The English man pages are moved back up one directory to mark their
status as the master copy for all translations. At the moment the po4a
translation infrastructure is defunct since it isn't integrated into
the automake environment yet. So far it hasn't been used, so that is
not an actual regression.

The date in the man pages is not automatically updated anymore. This is
as it should be, since the date is supposed to signify the time of the
last nontrivial change and not the release date of the software.

Signed-off-by: Andreas Bombe <aeb@debian.org>