set_fat(): Move FAT12 next cluster check up
In FAT12 two 12 bit entries are combined to a 24 bit value (three
bytes). Therefore, when an even numbered FAT entry is set in FAT12, it
must be be combined with the following entry. To prevent accessing
beyond the end of the FAT array, it must be checked that the cluster is
not the last one.

This check was broken in ff1b24e9 (first included in 3.0.3) as the
lookup was done unconditionally and the check influenced only using the
looked up value.

Move the check up to fix.

Signed-off-by: Andreas Bombe <aeb@debian.org>