Use just device size not major number in Atari mode
In Atari mode, read_boot() in boot.c used the device major number to
determine whether to use FAT12. It would always use FAT12 for a floppy,
otherwise only if it is a RAM disk or loopback device and has a size
corresponding to standard floppy formats.

Since this check was already broken for a long time (another place that
assumed 8 bit major numbers) and there is no real point to make the
distinction based on device, this commit reduces the check to just
compare against standard floppy sizes.

Signed-off-by: Andreas Bombe <aeb@debian.org>