Remove use of libblkid again
It appears libblkid is not as widely available as presumed, since some
platforms only have the original libblkid included in e2fsprogs which
lacks the needed functionality. This commit removes the requirement and
use of libblkid.

As a replacement, blkdev.c from util-linux is included, which offers the
required basic functionality in a portable way.

Signed-off-by: Andreas Bombe <aeb@debian.org>