Remove name/extension split in directory entry structures
Both the DIR_ENT structure in fsck.fat.h and the msdos_dir_entry in
msdos_fs.h - these represent the on disk format of directory entries -
had the name field split into name[8] followed by ext[3].

By far the most operations on name are on the full name including
extension and they treated the name field as an 11 byte array. This is
an array overflow that worked because the structs have the attribute
packed and the extension field is following right after.

Nevertheless, this is not clean C and the merging of both fields
actually simplified the code in a few places.

Signed-off-by: Andreas Bombe <aeb@debian.org>