mkfs: Fix off-by-2 error in bad cluster marking
mark_FAT_sector(), which has the mark_sector_bad() macro as its sole
user, computed the cluster number corresponding to the sector by taking
its offset from the first data sector and dividing by sectors per
cluster.

What it missed was that the first data cluster is number 2 and not 0.
This meant all marks were off by 2 and when the first two clusters are
supposed to be marked, it would overwrite the reserved cluster values
and create an invalid filesystem.

Signed-off-by: Andreas Bombe <aeb@debian.org>