mkfs: Fix offset error in FAT12/16 bad cluster marking
The root directory wasn't factored in to the calculation of the data
area start sector. On FAT32 the root directory is in the data area, but
for FAT12 and FAT16 it is a reserved space before the start of the data
area.

On FAT12 and FAT16, this resulted in the wrong clusters being marked
during bad blocks mapping, whether from check_blocks() or from reading
the user supplied bad blocks file.

Signed-off-by: Andreas Bombe <aeb@debian.org>