mkfs: Improved bounds checking in mark_FAT_sector()/mark_FAT_cluster()
In mark_FAT_sector() the sector number itself is now checked against
limits instead of the computed cluster number. Even with sector number
before the start of the data area, the cluster number may be valid for
the first cluster due to dividing by the cluster size.

Both functions now check for upper limits and should prevent writing
past the valid end of the FAT.

Signed-off-by: Andreas Bombe <aeb@debian.org>