mkfs: Limit filesystem size on targets that are too large
For FAT filesystems, the number of sectors has to fit into a 32 bit
variable. Previously this was not checked possibly causing invalid
filesystems to be generated.

Now there is a check for that case which will limit the number of
sectors if needed and print a warning that disk space will be left
unused in that case.

Signed-off-by: Andreas Bombe <aeb@debian.org>