mkfs: Improve parsing of bad blocks file
The bad blocks file that can be given to mkfs via the -l option had a
very simplistic design. It failed to notice it was parsing an empty
line and would report errors for that.

Replace it with a more robust version that ignores empty lines as well
as leading and trailing white space. Additionally it produces meaningful
error messages.

GitHub: Fixes #17
Signed-off-by: Andreas Bombe <aeb@debian.org>