Reinstate alignment of FAT32 structures to cluster size
This reverts commits 17c956cb9 and d63e0d627 where the alignment was
removed because it created problems with a device that refused to read
the aligned filesystem. The option -a is already provided to disable
alignment in order to handle such cases.

This change brings it back in line with FAT12/16 where alignment wasn't
disabled and brings consistency with the current command line options,
where only the option to disable alignment exists but no opposite option
to enable it.

Signed-off-by: Andreas Bombe <aeb@debian.org>