Add fsck testsuite
Add the framework for a fsck testsuite and a few tests, including a test
for the yet unimplemented check for misordered dot/dotdot directory
entries.

The test driver runs fsck twice on a given image and fails the test if
the first run does not detect an error or the second run still detects
an error.

Signed-off-by: Andreas Bombe <aeb@debian.org>