Fix definition and initialization of fsinfo sector
Change the definition of the fsinfo struct to be more in line with the
official definition, in particular make the trailing signature field 32
bit instead of 16 bit. During initialization in init_fsinfo(), first
clear the whole struct to zero before setting fields, to make sure
reserved fields are zeroed.

Fixes #33.

Signed-off-by: Andreas Bombe <aeb@debian.org>