Fix octal character parsing in file_cvt()
Apparently the octal code parsing got rewritten into a loop at some
point while the rest of the code was not removed. It was impossible to
specify an octal character code without the function returning an error.

This affected the -u and -d options for fsck.

Found through the dead code detection of Coverity Scan.

Signed-off-by: Andreas Bombe <aeb@debian.org>