mkfs: Remove redundant unreachable warning message
At some point the check for too many clusters for user requested FAT16
was moved a few lines above. The original location can only be reached
when the cluster count is too low for FAT16 which means the too many
clusters message can not be triggered there.

Found through the dead code detection of Coverity Scan.

Signed-off-by: Andreas Bombe <aeb@debian.org>