Request minimum automake version 1.11 and add workaround for <1.13
Before automake 1.13, parallel-tests wasn't the default set it
explicitly in the automake options list. The parallel test harness was
introduced in 1.11 therefore that is the minimum now.

However, the AM_TESTS_ENVIRONMENT variable was only starting to work in
1.13, so there is now a version check and a conditional assignment to
TESTS_ENVIRONMENT (which should be for user setting only) in case
automake is 1.11 or 1.12.

Without it, the fact of xxd being found is not passed to the test
scripts which then skip all tests due to perceived lack of xxd.

Signed-off-by: Andreas Bombe <aeb@debian.org>