fsck: Fix reporting of length to truncate with shared clusters
When files share clusters and one of the two files has been selected to
truncate fsck reports which file has been selected and the length to
which it will be truncated. All code paths erronously printed the length
of the file that was not going to be truncated, leading to possible user
confusion.

Signed-off-by: Andreas Bombe <aeb@debian.org>