fatlabel: Make sure printf style patterns in label are not expanded
Writing a label to an unlabeled filesystem involves creating a root
directory entry to contain the label. For this, the
alloc_rootdir_entry() function used for creating new directory nodes for
recovered orphan cluster chains is reused. However, this expects the
filename to be a printf style pattern to generate a unique name.

If the filesystem is unlabeled and the new label contains a printf
specifier other than %d or similar, the attempted expansion of the
pattern might crash fatlabel.

Change alloc_rootdir_entry() to take an additional argument gen_name and
only expand the name if its value is true.

Signed-off-by: Andreas Bombe <aeb@debian.org>