Add get_choice() function
The get_choice() function is intended to supersede get_key(). It takes
pairs of return values and strings for every choice and a
non-interactive default choice, does the prompting and reading of user
input (if interactive) and returns the value for the selected choice.

Since it needs to know if it runs in interactive mode, the "interactive"
variable has been moved to common.c and removed from fsck.fat.c and
fatlabel.c.

Signed-off-by: Andreas Bombe <aeb@debian.org>