fsck: Add quit choice to interactive prompts
All prompts now get an additional 'q' choice to explicitly quit fsck.
Selecting 'q' will print a message detailing the situation and ask for
confirmation.

Since the message describing the situation depends on whether fsck is in
immediate-write mode, the variable selecting that is moved to common.c
to make it accessible to the get_choice() function.

Signed-off-by: Andreas Bombe <aeb@debian.org>