mkfs: Link with common.c and use its die() implementation
Now common.c is linked into mkfs and its die() implementation is used
instead of the local implementation. That implementation consisting of a
die() macro and fatal_error() function has been removed.

To reproduce current messages exactly, there is now a program_name
variable in common.c. If it is not NULL, die() will prepend its contents
to the message it prints.

The previous implementation implicitly passed device_name to the
printf() so that die() invocations only needed to put "%s" where they
wanted the device name to appear in the string. These invocations now
have explicit device_name arguments.

Signed-off-by: Andreas Bombe <aeb@debian.org>