manpages, configure.ac: Correct description of Atari format selection
The -A option doesn't turn on (or turn off) Atari format selection, it
toggles it with every occurrence. Change wording in the man pages to
reflect that.

Also mention that automatic selection can happen only on 68k Atari Linux
in the man pages and the --enable-atari-check option help in
configure.ac.

Signed-off-by: Andreas Bombe <aeb@debian.org>