Remove leftover handling of split directory entry name
The name field in DIR_ENT used to be split into name and ext. Reading
the volume label from a DIR_ENT used to require merging the contents of
both fields and that handling was still there as reading parts of the
same name field and merging them. Simply read in one go instead.

Signed-off-by: Andreas Bombe <aeb@debian.org>