Harmonize unknown command line option handling
Now all binaries will show help and not try to print something like
"unknown option '?'" because '?' is the value getopt_long() returns when
it encounters an unknown option. At this point, getopt_long() has
already printed the problem to stderr so we don't have to.

The default action on getopt_long() return values (neither known option
nor '?') is to report an internal error with the unexpected value.

Signed-off-by: Andreas Bombe <aeb@debian.org>