Initialize iconv "from codepage" with //TRANSLIT
This would allow to show characters from DOS codepage on systems which
current locale does not support it via transliteration.

For example byte E8 in DOS CP437 (e with a grave accent) does not have
representation in Linux locale Latin2. But it can be shown after
transliteration as simple character "e".