Fix allowing changes in SFNs reporting bogus error on FAT32 root directory
The code to permit spaces in short filenames introduced in commit
4d01db7c6 caused the root directory pseudo file to be checked for a bad
name. Since the name for that file is not set (zeroed out), bad_name()
would report an error. Attempting to fix it would do nothing as the
rename functions check for and ignore the root directory.

Fix by checking for the root directory in bad_name().