fsck.fat: Reduce duplicate file name lookups
If a file has to be generated, then alloc_rootdir_entry tries to create
a unique name by using a counter curr_num for name generation and
looking for duplicates. If no duplicate is found, the counter curr_num
is not incremented, which means that the next file uses the same name
again. It is only incremented after the previously created file is
detected, just to repeat the loop again.

The improvement is simple: Increment the counter after creating the
filename. This has also the benefit that the check and use of
curr_num are closer to each other within the source code.

Signed-off-by: Samanta Navarro <ferivoz@riseup.net>