mkfs.fat: Fix signal race
The signal handler used during bad block detection calls the
signal-unsafe function printf. This could get in conflict with a printf
for an error condition in do_check.

Just set a sig_atomic_t variable to indicate that current status should
be displayed and leave the signal handler again.

The main loop checks the flag and prints the status when needed.

Signed-off-by: Samanta Navarro <ferivoz@riseup.net>