mkfs.fat: Switch to sigaction
According to signal(2) manual page the system call is only portable when
setting a signal's disposition to SIG_DFL or SIG_IGN.

Using sigaction(2) allows specific definition of flags, which in this
case will always instruct system calls to restart when SIGALRM is
received. This should be the default on Linux with signal(2) anyway.

This reduces the signal handler to set only a single variable.

Signed-off-by: Samanta Navarro <ferivoz@riseup.net>